requirements.txt
# --- CORE FRAMEWORK ---
fastapi==0.127.1
uvicorn==0.40.0
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.12.0
mcp==1.25.0
click>=8.1.0

# --- NETWORK & API ---
httpx==0.28.1
requests>=2.31.0
python-multipart>=0.0.9
openai>=1.12.0

# --- LLM DRIVERS ---
# Google Gemini (L3)
google-genai==1.56.0
google-api-python-client>=2.100.0
google-auth>=2.20.0

# Ollama (L1)
ollama==0.6.1
anthropic>=0.18.0

# --- MEMORY & AI (THE FIX) ---
# CRITICAL: Force Numpy 1.x to prevent "RuntimeError" or "__version__" crashes
numpy==1.26.4

# PyTorch (Let it pull its own nvidia deps)
torch>=2.2.0 --index-url https://download.pytorch.org/whl/cu121

# Vector DB & Embeddings
chromadb>=0.4.24
sentence-transformers>=2.6.0
transformers>=4.39.0
huggingface-hub>=0.22.0
tokenizers>=0.15.0

# Gravitas Grounded Research (Hybrid Search)
qdrant-client==1.16.2
minio==7.2.20
FlagEmbedding>=1.2.0

# Database (Postgres)
asyncpg==0.31.0

# --- DEV TOOLS & TESTING ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.3.0
black>=24.2.0
coloredlogs>=15.0.0
# --- UTILITIES ---
gputil==1.4.0
pyjwt==2.10.1
